Move Gemma4 CLI runtime out of main
